# server/internal/api/swagger/ui/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File4Base API Engine",
    "description": "Enterprise-grade REST and Dynamic Relational DDL Database Engine for File4Base. Supports runtime schema migrations, table occurrences, relationships graph, dynamic Find Mode formula parsing, security ACL, and solution archiving. Every endpoint except the health probes, sign-in and the database selector requires a session token: sign in with POST /api/v1/auth/login and send the returned token as 'Authorization: Bearer <token>'. A session is bound to the database it signed in to.",
    "version": "0.5.0",
    "contact": {
      "name": "File4Base Core Team",
      "url": "https://github.com/file4base/file4base-app"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local API Service (Direct)"
    },
    {
      "url": "http://localhost:3000",
      "description": "WebDirect Nginx Reverse Proxy"
    },
    {
      "url": "/",
      "description": "Current Host"
    }
  ],
...
  "paths": {
...
    "/api/v1/data/{table}": {
      "get": {
        "tags": [
          "Data Operations"
        ],
        "summary": "List Records in Table",
        "description": "Queries records from a table with pagination and sorting support.",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_asc",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of row records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Insert Record",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Record inserted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/data/{table}/{id}": {
      "get": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Get Record by ID",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Record not found"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Update Record",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record updated successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Delete Record",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Record deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/data/{table}/find": {
      "post": {
        "tags": [
          "Data Operations"
        ],
        "summary": "Execute Find Mode Search",
        "description": "Executes a Find query supporting wildcards (*), single characters (@/?), comparisons (>, <, >=, <=), exact (=), strict (==), negation (! / !=), ranges (...), today's date (//), and empty/non-empty search.",
        "parameters": [
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindRequest"
              },
              "example": {
                "criteria": {
                  "first_name": "Mario*",
                  "age": "20...50"
                },
                "omit": false,
                "limit": 100
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matching found records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FindRequest": {
        "type": "object",
        "required": [
          "criteria"
        ],
        "properties": {
          "criteria": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "omit": {
            "type": "boolean",
            "default": false
          },
          "limit": {
            "type": "integer",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "default": 0
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 Problem Details for HTTP APIs",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Invalid column format"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/schemas/tables"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token returned by POST /api/v1/auth/login. Bound to one user and one database; expires after a period of inactivity (SESSION_TTL, default 12h)."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid or expired session token",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The session's role or layout permissions do not allow this operation",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}